Version 2.0.0.30
Added option to specify a second precision separator, to be used as
separation between MilliSec and NanoSec.
Added option to specify a custom error value for invalid timestamps, or
incorrectly decoded timestamps.
Added 2 sql's for database schema and import of data, along with
instructions for how to import the csv into a MySql database.